Fix: resolve redundant branch, deprecated convert call and unnecessary cast warnings in mapping modules

+++ mcl/src/commonMain/kotlin/me/voltual/mcl/mapping/modules/MclRedstoneMapping.kt
@@ -39,7 +39,7 @@ object MclRedstoneMapping : MclMappingModule {
                 FacingDirection.EAST -> "mcl_observers:observer" to 1
                 FacingDirection.WEST -> "mcl_observers:observer" to 3
             }
-            MclNode("$nodeBase$suffix", param2 = (param2 as Int).toByte())
+            MclNode("$nodeBase$suffix", param2 = param2.toByte())
         })
 
         // ==========================================
